# swagger.yaml
openapi: 3.0.3
info:
  title: Girls Got Game API
  description: Basketball training application API for mobile and web clients
  version: 2.0.0
  contact:
    name: Girls Got Game API Support
    email: support@girlsgotgame.app
servers:
  - url: http://localhost:3001/api
    description: Development server
  - url: https://api.girlsgotgame.app/api
    description: Production server
security:
  - cookieAuth: []
paths:
...
  /mobile-auth/sign-in/mobile:
    post:
      tags:
        - Mobile Authentication
      summary: Mobile OAuth sign-in
      description: Authenticate mobile users with Google ID token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                provider:
                  type: string
                  enum: [google]
                  example: google
                idToken:
                  type: string
                  description: Google ID token from mobile OAuth
              required:
                - provider
                - idToken
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/AuthUser'
                  session:
                    $ref: '#/components/schemas/Session'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /mobile-auth/sign-out/mobile:
    post:
      tags:
        - Mobile Authentication
      summary: Mobile sign out
      description: Sign out mobile user and invalidate session
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Successfully signed out
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '500':
          $ref: '#/components/responses/InternalServerError'

  # Profile Endpoints
...
  /games:
    get:
      tags:
        - Games
      summary: Get all games
      description: Retrieve all games (public endpoint)
      responses:
        '200':
          description: Games retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Game'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Games
      summary: Create game
      description: Create a new game (admin only)
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGame'
      responses:
        '201':
          description: Game created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /games/{gameId}:
    get:
      tags:
        - Games
      summary: Get game by ID
      description: Retrieve a specific game by ID
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Game retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameWithDetails'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    patch:
      tags:
        - Games
      summary: Update game
      description: Update an existing game (admin only)
      security:
        - cookieAuth: []
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGame'
      responses:
        '200':
          description: Game updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Games
      summary: Delete game
      description: Delete a game (admin only)
      security:
        - cookieAuth: []
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Game deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /games/{gameId}/activities:
    get:
      tags:
        - Games
      summary: Get game activities
      description: Retrieve activities for a specific game
      security:
        - cookieAuth: []
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Game activities retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GameActivity'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /games/user/{userId}:
    get:
      tags:
        - Games
      summary: Get games for specific user
      description: Retrieve games where a specific user participated (admin or self only)
      security:
        - cookieAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User games retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Game'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # Admin Endpoints
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
    AuthUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
          nullable: true
        isAdmin:
          type: boolean
          default: false
    Session:
      type: object
      properties:
        id:
          type: string
        expiresAt:
          type: string
          format: date-time
        token:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        userId:
          type: string
...
    Game:
      type: object
      properties:
        id:
          type: string
          format: uuid
        teamName:
          type: string
        isHome:
          type: boolean
        opponentTeam:
          type: string
        gameDate:
          type: string
          format: date-time
        homeScore:
          type: integer
          nullable: true
        awayScore:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
        status:
          type: string
          enum: [scheduled, in_progress, completed, cancelled]
        isSharedToFeed:
          type: boolean
          default: false
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    GameWithDetails:
      allOf:
        - $ref: '#/components/schemas/Game'
        - type: object
          properties:
            players:
              type: array
              items:
                $ref: '#/components/schemas/GamePlayer'
            activities:
              type: array
              items:
                $ref: '#/components/schemas/GameActivity'
    GamePlayer:
      type: object
      properties:
        id:
          type: string
          format: uuid
        gameId:
          type: string
          format: uuid
        userId:
          type: string
          nullable: true
        manualPlayerId:
          type: string
          format: uuid
          nullable: true
        playerName:
          type: string
        jerseyNumber:
          type: integer
          nullable: true
    GameActivity:
      type: object
      properties:
        id:
          type: string
          format: uuid
        gameId:
          type: string
          format: uuid
        activityType:
          type: string
        description:
          type: string
        metadata:
          type: object
          nullable: true
        createdAt:
          type: string
          format: date-time
    CreateGame:
      type: object
      required:
        - teamName
        - isHome
        - opponentTeam
        - gameDate
      properties:
        teamName:
          type: string
        isHome:
          type: boolean
        opponentTeam:
          type: string
        gameDate:
          type: string
          format: date-time
        notes:
          type: string
    UpdateGame:
      type: object
      properties:
        teamName:
          type: string
        isHome:
          type: boolean
        opponentTeam:
          type: string
        gameDate:
          type: string
          format: date-time
        homeScore:
          type: integer
        awayScore:
          type: integer
        notes:
          type: string
        status:
          type: string
          enum: [scheduled, in_progress, completed, cancelled]
        isSharedToFeed:
          type: boolean
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Invalid data provided
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Authentication required
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Insufficient permissions
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Resource not found
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Internal server error
...